gh-634 Move TestDataGenerator to stroom-util, add xml handling
