+++ stroom-util/src/test/java/stroom/util/test/data/DataWriter.java
@@ -1,0 +1,9 @@
+package stroom.util.test.data;
+
+import java.util.List;
+import java.util.stream.Stream;
+
+public interface DataWriter {
+    Stream<String> mapRecords(final List<Field> fieldDefinitions,
+                              final Stream<Record> recordStream);
+}
